changed gender icons to be business appropriate

+++ app/src/main/java/com/team16/sopra/sopra16team16/Controller/ContactManager.java
@@ -381,8 +381,10 @@ public class ContactManager{
 
         File[] files = filesPath.listFiles();
 
-        for (int i = 1; i < files.length; i++) {
-            FileUtils.deleteFile(files[i].getPath());
+        if (files != null) {
+            for (int i = 1; i < files.length; i++) {
+                FileUtils.deleteFile(files[i].getPath());
+            }
         }
     }
 }
